flask_server/requirements.txt
Flask==2.0.2
requests==2.27.1
gunicorn
numpy
Pillow